More tidy ups
